✨ feat: 保存json到指定路径

+++ main.ts
@@ -8,6 +8,7 @@ import boxen from "boxen";
 import server from "./src/server";
 import { Server, IncomingMessage, ServerResponse } from "http";
 import { runAnalysis } from "./src/analysis";
+import { exec } from "child_process";
 
 // 定义基本的url
 const baseUrl: string = path.join(__dirname, "src", "data");
@@ -24,31 +25,43 @@ program
   .option("--json [file-path]", "将依赖关系以json格式导入指定路径中")
   .action(async (options): Promise<void> => {
     // 获取深度
-    const depth = options.depth || 0;
+    const depth = options.depth || 9999;
     // 获取json文件路径
-    const jsonFilePath = options.json;
-    // // 执行依赖分析
-    await runAnalysis(depth);
-    // // 启动服务器
-    const serverInstance: Server<
-      typeof IncomingMessage,
-      typeof ServerResponse
-    > = server.listen(port, () => {
-      console.log(getBeautifulMsg());
-    });
-    // 监听 Ctrl+C 退出事件
-    process.on("SIGINT", () => {
-      // 关闭express
-      serverInstance.close();
-      // 删除临时文件
-      const files: string[] = fs.readdirSync(baseUrl);
-      files.forEach((item) => {
-        const filePath: string = path.join(baseUrl, item);
-        fs.unlinkSync(filePath);
+    const jsonFilePath = options.json || 'default';
+    // 
+    if (depth === true || jsonFilePath === true) {
+      console.error('请输入正确的参数');
+      return;
+    }
+    // 执行依赖分析
+    await runAnalysis(depth, jsonFilePath);
+    // 如果输入了 jsonFilePath 就打开对应目录
+    if (jsonFilePath !== 'default') {
+      exec(`start ${jsonFilePath}`, (err) => {
+        if (err) { console.error('Failed to save:', err); return; }
+      })
+    } else {
+      // 启动服务器
+      const serverInstance: Server<
+        typeof IncomingMessage,
+        typeof ServerResponse
+      > = server.listen(port, () => {
+        console.log(getBeautifulMsg());
       });
-      console.log("Bye~");
-      process.exit(0);
-    });
+      // 监听 Ctrl+C 退出事件
+      process.on("SIGINT", () => {
+        // 关闭express
+        serverInstance.close();
+        // 删除临时文件
+        const files: string[] = fs.readdirSync(baseUrl);
+        files.forEach((item) => {
+          const filePath: string = path.join(baseUrl, item);
+          fs.unlinkSync(filePath);
+        });
+        console.log("Bye~");
+        process.exit(0);
+      });
+    }
   });
 
 // 获取消息
